back_end/requirements.txt
asgiref==3.4.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
defusedxml==0.7.1
Django==4.0
django-cors-headers==3.10.1
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
httpie==2.6.0
idna==3.3
install==1.3.5
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.2
Pygments==2.10.0
PyJWT==2.3.0
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
